requirements.txt
streamlit
scikit-learn
numpy
pandas
openai
python-dotenv